zss-app/backend/sso_service/requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
python-dotenv==0.21.0
bcrypt==4.0.1
PyJWT==2.6.0
Flask-Cors==3.0.10